requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
gunicorn==19.9.0